Slight modifications in the locations of various files

+++ preprocessingCodes/saveRawData.m
@@ -1,19 +1,23 @@
 % saveRawData
 % The raw data obtained during the experiment is copied to an appropriate folder.
 
-pairedSubjectNameList = getMatchedSubjectsBK1;
+tmp = load('BK1AllSubjectList.mat');
+allSubjectList = tmp.allSubjectList;
+
 [subjectNameList,expDateList] = getDemographicDetails('BK1');
 
 folderSourceString = 'N:\Projects\MeditationProjects\MeditationProject2\data\rawData';
 folderDestinationString = 'N:\Projects\ProjectDhyaan\BK1\data\rawData';
 
-for i=1:length(pairedSubjectNameList)
-    for j=1:2
-        subjectName = pairedSubjectNameList{i,j};
-        expDate = expDateList{strcmp(subjectName,subjectNameList)};
-        folderIn = fullfile(folderSourceString,[subjectName expDate]);
-        folderOut = fullfile(folderDestinationString,[subjectName expDate]);
+for i=1:length(allSubjectList)
+    subjectName = allSubjectList{i};
+    expDate = expDateList{strcmp(subjectName,subjectNameList)};
+    folderIn = fullfile(folderSourceString,[subjectName expDate]);
+    folderOut = fullfile(folderDestinationString,[subjectName expDate]);
 
+    if exist(folderOut,'file')
+        disp([folderOut ' exists']);
+    else
         disp(['Copying ' folderIn ' to ' folderOut]);
         copyfile(folderIn,folderOut,'f');
     end
